feat: move get config function into modules.py to avoid circular imports
