+++ recognition/siamese_s46387334/predict.py
@@ -18,7 +18,6 @@ import numpy as np
 
 from dataset import get_isic2020_data, get_isic2020_data_loaders
 from modules import SiameseNet, set_seed
-from train import get_config
 
 
 ###############################################################################
